+++ CHANGELOG.md
@@ -41,6 +41,8 @@ Recorder.FileType changed from method to write-only property
 - Recorder.State
 - Command.Reset()
 - Command.Save($filepath)
+- Command.StorePreset()
+- Command.RecallPreset()
 
 ### Changed